Move types to module Network.Mail.SMTP.Types
I did not move SMTPConnection, as it is the only non-pure type (uses Handle).
I plan to move the protocol implementation into an "Abstract" module, which
will only use pure types.
